Reset header flag after writing summary
The missing reset of the "resultHeaderWritten" flag causes missing
column headers in the result file, if the ResultCollector is called
multiple times for different analyses, like in test cases. This flag
will now be reset every time after the summary file is written, which
marks the end of an analysis.