Fix OutOfBoundsException
That exception occurs in rare cases where "depends on" is changed and
the first line of the definition of a configuration option has a
whitespace. In that case the "isPartOfHelp"-check was searching for the
start of a help or comment text potentially starting at index -1.
That is fixes by terminating further search, if the given start index is
already 0.